Frontend/openai_utils.py
import os
from openai import OpenAI
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Get API key
api_key = os.getenv("OPENAI_API_KEY")

if not api_key:
    raise ValueError("OPENAI_API_KEY not found in .env file")

# Create OpenAI client
client = OpenAI(api_key=api_key)


def ask_openai(prompt: str):
    """
    Sends a prompt to OpenAI and returns a short,
    human-like response (2–3 lines).
    """

    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {
                "role": "system",
                "content": "You are a professional but human-like AI assistant. Keep answers concise and clear."
            },
            {
                "role": "user",
                "content": prompt
            }
        ],
        temperature=0.7,
        max_tokens=120
    )

    return response.choices[0].message.content.strip()
